requirements.txt
Babel==2.12.1
folium==0.14.0
inflection==0.5.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
streamlit==1.22.0
streamlit_folium==0.11.1
